Falta lang

+++ database/migrations/2021_02_20_000001_create_orders_table.php
@@ -15,7 +15,7 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('total');
+            $table->double('total');
             $table->bigInteger('user')->unsigned();
             $table->foreign('user')->references('id')->on('users');
             $table->date('date');
